Tool tip merged, Added specific teams colour's, removed halo from units, Minor fixes

+++ Assemble Z' Army - Project 2022/Assets/Scripts/Combat/SwordAttack.cs
@@ -17,7 +17,7 @@ public class SwordAttack : Attacker
     {
         GetComponent<Animator>().SetBool("isRunning", false);
         GetComponent<Animator>().SetBool("isAttacking", true);
-        target.GetComponent<Health>().DealDamage(damage);
+        target.GetComponent<Health>().DealDamage((int)damage);
         print("Recruit should attack");
     }
